+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## 0.1.0 (0000-00-00)
 
--
+- New Functions
 
+  - Disable-CorporateUser